refactor(utils): add NAV_LINKS to utils
