+++ src/utils/navLinks.ts
@@ -1,0 +1,10 @@
+export const NAV_LINKS = [
+  {
+    link: "/cellars",
+    title: "Cellars",
+  },
+  {
+    link: "/bridge",
+    title: "Bridge",
+  },
+]
